Conflicts:
	webapp/WEB-INF/jsp/user/browser.jsp

+++ src/com/epam/lab/controller/services/user/UserServiceImpl.java
@@ -7,6 +7,7 @@ import java.util.List;
 
 import org.apache.log4j.Logger;
 
+import com.epam.lab.controller.annotations.TableColumn;
 import com.epam.lab.controller.dao.folder.FolderDAOImpl;
 import com.epam.lab.controller.dao.tariff.TariffDAOImpl;
 import com.epam.lab.controller.dao.token.TokenDAOImpl;
@@ -37,7 +38,7 @@ public class UserServiceImpl extends AbstractServiceImpl<User> implements
 	public UserServiceImpl() {
 		super(new UserDAOImpl());
 	}
-	
+
 	public long getCount() {
 		return userDaoImpl.getCount();
 	}
@@ -53,7 +54,7 @@ public class UserServiceImpl extends AbstractServiceImpl<User> implements
 		user.setIsBanned(false);
 		insert(user);
 	}
-	
+
 	public List<User> getBySQL(String sql) {
 		return userDaoImpl.getBySQL(sql);
 	}
@@ -92,7 +93,7 @@ public class UserServiceImpl extends AbstractServiceImpl<User> implements
 	public void deleteFilesAndFolders(String[] filesId, String[] foldersId) {
 		long[] files = null;
 		long[] folders = null;
-		if(filesId !=null){
+		if (filesId != null) {
 			files = new long[filesId.length];
 			for (int i = 0; i < filesId.length; i++) {
 				try {
@@ -102,7 +103,7 @@ public class UserServiceImpl extends AbstractServiceImpl<User> implements
 				}
 			}
 		}
-		if(foldersId !=null && !foldersId[0].equals("")){
+		if (foldersId != null && !foldersId[0].equals("")) {
 			folders = new long[foldersId.length];
 			for (int i = 0; i < foldersId.length; i++) {
 				try {
@@ -131,10 +132,13 @@ public class UserServiceImpl extends AbstractServiceImpl<User> implements
 	}
 
 	@Override
-	public void deleteUsers(String[] usersId) {
+	public void deleteUsers(String[] usersId, Long idAdmin) {
 		UserFileServiceImpl fileService = new UserFileServiceImpl();
 		FolderServiceImpl folderService = new FolderServiceImpl();
 		for (int i = 0; i < usersId.length; i++) {
+			if (Long.parseLong(usersId[i]) == idAdmin) {
+				continue;
+			}
 			fileService.deleteByUserId(Long.parseLong(usersId[i]));
 			folderService.deleteByUserId(Long.parseLong(usersId[i]));
 			userDaoImpl.delete((Long.parseLong(usersId[i])));
@@ -148,9 +152,12 @@ public class UserServiceImpl extends AbstractServiceImpl<User> implements
 	}
 
 	@Override
-	public void activateUsers(String[] usersId) {
+	public void activateUsers(String[] usersId, Long idAdmin) {
 		UserDAOImpl userDaoImpl = new UserDAOImpl();
 		for (int i = 0; i < usersId.length; i++) {
+			if (Long.parseLong(usersId[i]) == idAdmin) {
+				continue;
+			}
 			userDaoImpl.setIsActivate(true, Long.parseLong(usersId[i]));
 			MailSender.send(get(Long.parseLong(usersId[i])).getEmail(),
 					"Activation", "You're activated!!!");
@@ -158,8 +165,12 @@ public class UserServiceImpl extends AbstractServiceImpl<User> implements
 	}
 
 	@Override
-	public void banedUsers(String[] usersId) {
+	public void banedUsers(String[] usersId, Long idAdmin) {
+		System.out.println(idAdmin);
 		for (int i = 0; i < usersId.length; i++) {
+			if (Long.parseLong(usersId[i]) == idAdmin) {
+				continue;
+			}
 			userDaoImpl.setIsBanned(true, Long.parseLong(usersId[i]));
 			MailSender.send(get(Long.parseLong(usersId[i])).getEmail(), "Ban",
 					"You're baned!!!");
@@ -167,8 +178,11 @@ public class UserServiceImpl extends AbstractServiceImpl<User> implements
 	}
 
 	@Override
-	public void cancelBanUsers(String[] usersId) {
+	public void cancelBanUsers(String[] usersId, Long idAdmin) {
 		for (int i = 0; i < usersId.length; i++) {
+			if (Long.parseLong(usersId[i]) == idAdmin) {
+				continue;
+			}
 			userDaoImpl.setIsBanned(false, Long.parseLong(usersId[i]));
 			MailSender.send(get(Long.parseLong(usersId[i])).getEmail(),
 					"Cancel ban", "You're not baned!!!");
@@ -176,30 +190,25 @@ public class UserServiceImpl extends AbstractServiceImpl<User> implements
 
 	}
 
-	public void changeUserPassword(User user, String password) {
+	public void changeUserPassword(long userId, String password) {
 		MD5Encrypter md5 = new MD5Encrypter();
 		String md5Pass = md5.encrypt(password);
 		UserDAOImpl userDaoImpl = new UserDAOImpl();
+		User user = get(userId);
 
 		user.setPassword(md5Pass);
 		userDaoImpl.update(user);
 	}
 
-	public String sendUsersEmail(String[] usersId) {
+	public void sendUsersEmail(String[] usersId, String subject, String message) {
 		UserDAOImpl userDaoImpl = new UserDAOImpl();
-		String errorMessage = null;
 		List<User> users = new ArrayList<User>();
-		if (usersId == null) {
-			errorMessage = "Please check the users you want to send email!!!";
-		} else {
-			for (int i = 0; i < usersId.length; i++) {
-				users.add(userDaoImpl.get(Long.parseLong(usersId[i])));
-			}
-			for (User iter : users) {
-				MailSender.send(iter.getEmail(), "dreamhost", "test message");
-			}
+		for (int i = 0; i < usersId.length; i++) {
+			users.add(userDaoImpl.get(Long.parseLong(usersId[i])));
 		}
-		return errorMessage;
+		for (User iter : users) {
+			MailSender.send(iter.getEmail(), subject, message);
+		}
 	}
 
 	public User changeUserLogin(String email, String login) {
@@ -371,7 +380,8 @@ public class UserServiceImpl extends AbstractServiceImpl<User> implements
 		return userDaoImpl.get(userId).getIsBanned();
 	}
 
-	public void moveFilesAndFolders(String[] moveAble, long idTarget, long userId) {
+	public void moveFilesAndFolders(String[] moveAble, long idTarget,
+			long userId) {
 		for (String move : moveAble) {
 			String type = move.split("-")[0];
 			long id = Long.parseLong(move.split("-")[1]);
@@ -388,28 +398,31 @@ public class UserServiceImpl extends AbstractServiceImpl<User> implements
 			}
 		}
 	}
-	
-	public List<User> getByParam(String page, String count, String orderBy, String sop) {
+
+	public List<User> getByParam(String page, String count, String orderBy,
+			String sop) {
 		UserDAOImpl userDAO = new UserDAOImpl();
 		int p = getPage(page);
 		int c = getCount(count);
 		String order = getOrderBy(orderBy);
 		String sort = getSort(sop);
-		
+
 		StringBuilder sql = new StringBuilder();
-		
+
 		sql.append("SELECT * FROM ").append("users ");
 		sql.append(" ORDER BY ").append(order);
 		sql.append(" ").append(sort);
 		sql.append(" LIMIT ").append(c);
-		sql.append(" OFFSET ").append(p*c);
-		
+		sql.append(" OFFSET ").append(p * c);
+
+		System.out.println(sql.toString());
+
 		return userDAO.getBySQL(sql.toString());
 	}
-	
+
 	private int getPage(String page) {
 		int p = 0;
-		
+
 		if (page != null) {
 			try {
 				p = Integer.valueOf(page);
@@ -418,13 +431,13 @@ public class UserServiceImpl extends AbstractServiceImpl<User> implements
 				p = 0;
 			}
 		}
-		
+
 		return p;
 	}
-	
+
 	private int getCount(String count) {
 		int c = 10;
-		
+
 		if (count != null) {
 			try {
 				c = Integer.valueOf(count);
@@ -433,46 +446,53 @@ public class UserServiceImpl extends AbstractServiceImpl<User> implements
 				c = 10;
 			}
 		}
-		
+
 		return c;
 	}
-	
+
 	private String getOrderBy(String orderBy) {
 		String order = null;
 		Field[] fields = User.class.getDeclaredFields();
-		
-		for (int i = 0; i < fields.length; i++) {
-			if (fields[i].getName().equals(orderBy)) {
-				order = orderBy;
-				break;
+
+		if (orderBy != null) {
+			for (int i = 0; i < fields.length; i++) {
+				if (fields[i].isAnnotationPresent(TableColumn.class)) {
+					TableColumn annotation = fields[i]
+							.getAnnotation(TableColumn.class);
+					if (annotation.value().equals(orderBy)) {
+						order = annotation.value();
+						break;
+					}
+				}
 			}
-		}
-		
-		if (order == null) {
+		} else {
 			order = "id";
 		}
-		
+
 		return order;
 	}
-	
+
 	private String getSort(String sop) {
-		String sort = null;
-		
-		if (sop == null || !sop.toLowerCase().equals("asc") || !sop.toLowerCase().equals("desc")) {
-			sort = "asc";
-		} else {
-			sort = sop;
+		String sort = "asc";
+
+		if (sop != null) {
+			if (sop.toLowerCase().equals("asc")) {
+				sort = "asc";
+			} else if (sop.toLowerCase().equals("desc")) {
+				sort = "desc";
+			}
 		}
 
 		return sort;
 	}
-	public long getFreeSize(long userId){
+
+	public long getFreeSize(long userId) {
 		UserDAOImpl dao = new UserDAOImpl();
 		User user = dao.get(userId);
 		TariffDAOImpl tariffDao = new TariffDAOImpl();
 		Tariff tarriff = tariffDao.get(user.getIdTariff());
 		long freeSize = tarriff.getMaxCapacity() - user.getCapacity();
 		return freeSize;
-		
+
 	}
 }
